Added the separation between shake-off/up
Separated the shake-off and shake-up probabilities.
Added a shakeupinty file (missing due to being too large) for the shake-up monopolar transitions.
The shake-off transitions are still in the satinty file.
Although the separated probabilities are in the shakeoff and shakeup files the shakeweights file still has the total probability for easier use in the program.
The file groundshakeupenergy has the creation energy and width for the new initial shake-up levels.
Various conversion scripts have been added to convert the MCDF files.
The previous conversion scripts have also been updated to uniformly convert the level labels by reading them from the single conversion performed while generating the files with the creation energy for each level.